+++ firebase-vertexai/src/main/kotlin/com/google/firebase/vertexai/type/FunctionDeclarations.kt
@@ -185,12 +185,22 @@ class Schema<T>(
   fun fromString(value: String?) = type.parse(value)
 
   companion object {
-    /** Registers a schema for an integer number */
+    /** Registers a schema for a 32 bit integer number */
     fun int(name: String, description: String) =
+      Schema<Int>(
+        name = name,
+        description = description,
+        format = "int32",
+        type = FunctionType.INTEGER,
+        nullable = false,
+      )
+
+    /** Registers a schema for a 64 bit integer number */
+    fun long(name: String, description: String) =
       Schema<Long>(
         name = name,
         description = description,
-        type = FunctionType.INTEGER,
+        type = FunctionType.LONG,
         nullable = false,
       )
 
@@ -202,7 +212,6 @@ class Schema<T>(
         type = FunctionType.STRING,
         nullable = false,
       )
-
     /** Registers a schema for a boolean */
     fun bool(name: String, description: String) =
       Schema<Boolean>(
@@ -213,7 +222,7 @@ class Schema<T>(
       )
 
     /** Registers a schema for a floating point number */
-    fun num(name: String, description: String) =
+    fun double(name: String, description: String) =
       Schema<Double>(
         name = name,
         description = description,