+++ firebase-vertexai/CHANGELOG.md
@@ -3,3 +3,5 @@
 * [changed] Renamed `GoogleGenerativeAIException` to `FirebaseVertexAIException`.
 * [changed] Updated the KDocs for various classes and functions.
 * [fixed] Fixed an issue with decoding JSON literals (#6028).
+* [changed] renamed `Schema.int` to `Schema.long` and added a 32 bit int type
+* [changed] renamed `Schema.num` to `Schema.double`